serverless-src/requirements.txt
Flask==1.0.2
numpy==1.15.4
Pillow==5.4.1
google_api_python_client==1.7.7
